src/python/coypu_kafka/requirements.txt
kafka-python
websocket-client
psycopg2-binary
